Added test model function for model testing
